Move operator to subdir to bind version scope also
This implies that opertor image is going to be tagged based on
the subtree hash of operator subdir, so that any changes in bundles
subdir won't spoil the tag semantics.

In other words this means that running `make images.all` over and
over doesn't update bundles each time, as there are no changes to
the operator image.